Use `ed25519` + `signature` interop crates
The `ed25519` crate and the `signature` crates provide trait-based
interoperability between Ed25519 libraries:

https://github.com/RustCrypto/signatures

This commit integrates the `ed25519::Signature` type, and changes the
existing `sign` and `verify` methods to use the `Signer` and `Verifier`
traits from the `signature` crate. Additionally, it replaces the current
error types with an opaque one from the `signature` crate.

This has the drawback of requiring the `Signer` and `Verifier` traits
are in scope in order to create signatures, but the benefit of
interoperable support for other Ed25519 signers and verifiers. Presently
there are a number of these provided through the Signatory framework,
along with support for HSM-backed Ed25519 signatures in the `yubihsm`
crate:

- https://github.com/tendermint/signatory
- https://github.com/tendermint/yubihsm-rs