Draft: New facility for exposing "bundled signatures".
Sometimes it's useful to save up one or more signatures and have
them verified later on.  For example, in Arti, we sometimes want to
parse a large document containing many internal signatures, but save
the verification of the signatures for another thread.  We can't
use precomputed hashes for this, since the protocol is already
specified.  Thus, our only choice for now is to carry around the
original message--either by copy or by reference--until we're
ready to verify the signature on it.

With this patch, ed25519-dalek exposes a new type, BundledSignature.
A BundledSignature contains the public key, the InternalSignature,
and the scalar `k` needed to verify the signature against the message.
To avoid code bloat and to reuse testing, it uses these objects
internally to implement signature verification.  (According to
`cargo bench`, there is no performance loss.)