Rename `ExpandedSecretKey::nonce` => `prefix`
The name "nonce" doesn't really capture the role of this value, which
unfortunately seems to resist a pithy name across all of the specs
including RFC8032, which refers to it as `h_b, ..., h_(2b-1)`. A nonce
is used once whereas this value is used repeatedly for every message
signed.

Its usage is a prefix concatenated onto the message to be hashed and
used to compute `r`, so given that, this commit renames it to `prefix`.